third_party/blink/web_tests/http/tests/websocket/close-before-handshake-response-expected.txt

Test if WebSocket fires error and close event when handshake is not received at all.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS close event
PASS wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
